+++ src/router/index.js
@@ -1,9 +1,8 @@
 import Vue from 'vue';
 import VueRouter from 'vue-router';
 
-Vue.use(VueRouter);
-
 export const createRouter = (routes) => {
+  Vue.use(VueRouter);
   return new VueRouter({
     mode: 'history',
     base: process.env.BASE_URL,