+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 ## v0.2.0
 
 - Feat: add GitHub generate mode.
+
+- Feat: allow theme generators run before framework.
+
+- Feat: `vue-router` and `vuex` only be imported when it has been used by theme.